feat: added React-Native app generation to boostrapper
